refactored into sections
